services/ml_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.12
psycopg2-binary==2.9.9
kafka-python==2.0.2
numpy==2.1.3
pillow==10.4.0
matplotlib==3.9.2
torch==2.4.1
torchvision==0.19.1
